fix: separated multiple components into different files from alert-box.tsx
