+++ src/components/alert-box/alert-action-button.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { Button } from 'native-base';
+
+import { AlertActionButtonProps } from './../../types/alert';
+
+import { useAlertStyles } from './alert.styles';
+
+export const AlertActionButton: React.FC<AlertActionButtonProps> = ({ label, color, onPress }) => {
+  const styles = useAlertStyles();
+  return (
+    <Button onPress={onPress} bgColor={color} _text={styles.buttonText} style={styles.button}>
+      {label}
+    </Button>
+  );
+};
